Refactored unittests and runActions logic removed from index
