+++ src/index.ts
@@ -1,7 +1,6 @@
-import { runAction } from "./actions";
+import { runActions } from "./actions";
 import { gitOperations } from "./git_operations";
 import { startup } from "./startup";
-import { getPriorityRange } from "./priority";
 import { loadConfig } from "./config_loader";
 import { printLogs } from "./utils";
 import { searchOutputForHints } from "./search_output";
@@ -17,35 +16,19 @@ export async function start(
 
 	await startup(Object.values(cnf.startup));
 
-	const gitOperationsPromises = [];
-	for (const projectObj of Object.values(cnf.projects)) {
-		gitOperationsPromises.push(gitOperations(cwd, projectObj));
-	}
+	const gitOperationsPromises = Object.values(cnf.projects).map((project) =>
+		gitOperations(cwd, project),
+	);
 	const logs = await Promise.all(
 		gitOperationsPromises.map((p) => p.catch((e) => e)),
 	);
 	printLogs(Object.keys(cnf.projects), logs);
 
-	const prioRange = getPriorityRange(Object.values(cnf.projects));
-
-	const stdoutBuffer: (GroupKey & Output)[] = [];
-	for (let i = prioRange.min; i < prioRange.max; i++) {
-		const runActionPromises = [];
-		for (const projectKey of Object.keys(cnf.projects)) {
-			runActionPromises.push(
-				runAction(
-					cwd,
-					cnf,
-					{ project: projectKey, action: actionToRun, group: groupToRun },
-					i,
-				),
-			);
-		}
-		const stdoutBufferPromises = await Promise.all(runActionPromises);
-		stdoutBufferPromises
-			.filter((p) => p)
-			.forEach((p) => stdoutBuffer.push(p as GroupKey & { stdout: string }));
-	}
-
+	const stdoutBuffer: (GroupKey & Output)[] = await runActions(
+		cnf,
+		cwd,
+		actionToRun,
+		groupToRun,
+	);
 	if (cnf.searchFor) searchOutputForHints(cnf.searchFor, stdoutBuffer);
 }
