+++ src/types/config.ts
@@ -1,18 +1,18 @@
 export type Action = {
-	hint: string | undefined;
+	hint?: string;
 };
 export type CmdAction = Action & { cmd: [string, ...string[]] };
 export type ShellAction = Action & { shell: string; script: string };
 
 export type ProjectAction = {
-	priority: number | undefined;
+	priority?: number;
 	groups: { [key: string]: [string, ...string[]] };
 };
 
 export type Project = {
 	remote: string;
 	default_branch: string;
-	priority: number | undefined;
+	priority?: number;
 	actions: { [key: string]: ProjectAction };
 };
